partial sync with mcx_utils.c from mcx-cuda

+++ src/mcx_utils.h
@@ -184,7 +184,7 @@ typedef struct MCXConfig{
 extern "C" {
 #endif
 void mcx_savedata(float *dat, int len, Config *cfg);
-void mcx_savenii(float *dat, int len, char* name, int type32bit, int outputformatid, Config *cfg);
+void mcx_savenii(float *dat, size_t len, char* name, int type32bit, int outputformatid, Config *cfg);
 void mcx_error(const int id,const char *msg,const char *file,const int linenum);
 void mcx_assess(const int id,const char *msg,const char *file,const int linenum);
 void mcx_cleargpuinfo(GPUInfo **gpuinfo);
